refactor: Move deposit and transfer modals into BankBalanceSection component
