+++ app/src/components/sections/BankView/BankBalanceSection.vue
@@ -27,7 +27,7 @@
             v-if="bankAddress"
             variant="secondary"
             class="flex items-center gap-2"
-            @click="$emit('open-deposit')"
+            @click="depositModal = true"
             data-test="deposit-button"
           >
             <PlusIcon class="w-5 h-5" />
@@ -37,7 +37,7 @@
             v-if="bankAddress"
             variant="secondary"
             class="flex items-center gap-2"
-            @click="$emit('open-transfer')"
+            @click="transferModal = true"
             data-test="transfer-button"
           >
             <ArrowsRightLeftIcon class="w-5 h-5" />
@@ -48,9 +48,42 @@
 
       <div class="text-sm text-gray-600 mt-4">
         Contract Address:
-        <span class="font-mono">{{ bankAddress }}</span>
+        <span class="font-mono" data-test="bank-address">{{ bankAddress }}</span>
       </div>
     </div>
+
+    <!-- Deposit Modal -->
+    <ModalComponent v-model="depositModal">
+      <DepositBankForm
+        v-if="depositModal"
+        @close-modal="() => (depositModal = false)"
+        @deposit="depositToBank"
+        :loading="
+          depositLoading ||
+          isConfirmingDeposit ||
+          isPendingApprove ||
+          isConfirmingApprove ||
+          tokenDepositLoading ||
+          isConfirmingTokenDeposit
+        "
+        :loading-text="loadingText"
+      />
+    </ModalComponent>
+
+    <!-- Transfer Modal -->
+    <ModalComponent v-model="transferModal">
+      <TransferFromBankForm
+        v-if="transferModal"
+        @close-modal="() => (transferModal = false)"
+        @transfer="transferFromBank"
+        @searchMembers="(input: string) => searchUsers({ name: '', address: input })"
+        :filteredMembers="foundUsers"
+        :loading="transferLoading || isConfirmingTransfer"
+        :bank-balance="teamBalance?.formatted || '0'"
+        :usdc-balance="formattedUsdcBalance || '0'"
+        service="Bank"
+      />
+    </ModalComponent>
   </div>
 </template>
 
@@ -58,26 +91,86 @@
 import { PlusIcon, ArrowsRightLeftIcon } from '@heroicons/vue/24/outline'
 import ButtonUI from '@/components/ButtonUI.vue'
 import { NETWORK, USDC_ADDRESS } from '@/constant'
-import { useBalance, useReadContract, useChainId } from '@wagmi/vue'
-import { computed, watch } from 'vue'
+import {
+  useBalance,
+  useReadContract,
+  useChainId,
+  useSendTransaction,
+  useWaitForTransactionReceipt,
+  useWriteContract
+} from '@wagmi/vue'
+import { computed, ref, watch } from 'vue'
 import type { Address } from 'viem'
 import ERC20ABI from '@/artifacts/abi/erc20.json'
 import { useToastStore } from '@/stores/useToastStore'
 import { log, parseError } from '@/utils'
+import ModalComponent from '@/components/ModalComponent.vue'
+import DepositBankForm from '@/components/forms/DepositBankForm.vue'
+import TransferFromBankForm from '@/components/forms/TransferFromBankForm.vue'
+import { useCustomFetch } from '@/composables/useCustomFetch'
+import { useUserDataStore } from '@/stores/user'
+import BankABI from '@/artifacts/abi/bank.json'
+import { readContract } from '@wagmi/core'
+import { config } from '@/wagmi.config'
+import { parseEther } from 'viem'
+import type { User } from '@/types'
 
 const props = defineProps<{
   bankAddress: Address | undefined
 }>()
 
 const emit = defineEmits<{
-  (e: 'open-deposit'): void
-  (e: 'open-transfer'): void
   (e: 'error'): void
   (e: 'balance-updated'): void
 }>()
 
-const { addErrorToast } = useToastStore()
+const { addErrorToast, addSuccessToast } = useToastStore()
+const userDataStore = useUserDataStore()
+const currentAddress = userDataStore.address
 const chainId = useChainId()
+
+// Add refs for modals and form data
+const depositModal = ref(false)
+const transferModal = ref(false)
+const foundUsers = ref<User[]>([])
+const depositAmount = ref('')
+
+// Contract interactions
+const { sendTransaction, isPending: depositLoading, data: depositHash } = useSendTransaction()
+
+const { isLoading: isConfirmingDeposit } = useWaitForTransactionReceipt({
+  hash: depositHash
+})
+
+const {
+  data: transferHash,
+  isPending: transferLoading,
+  writeContract: transfer
+} = useWriteContract()
+
+const { isLoading: isConfirmingTransfer } = useWaitForTransactionReceipt({
+  hash: transferHash
+})
+
+const {
+  writeContract: writeTokenDeposit,
+  isPending: tokenDepositLoading,
+  data: tokenDepositHash
+} = useWriteContract()
+
+const { isLoading: isConfirmingTokenDeposit } = useWaitForTransactionReceipt({
+  hash: tokenDepositHash
+})
+
+const {
+  writeContract: approve,
+  isPending: isPendingApprove,
+  data: approveHash
+} = useWriteContract()
+
+const { isLoading: isConfirmingApprove } = useWaitForTransactionReceipt({
+  hash: approveHash
+})
 
 // Balance fetching
 const {
@@ -103,12 +196,114 @@ const {
   args: [props.bankAddress as Address]
 })
 
+// Functions
+const depositToBank = async (data: { amount: string; token: string }) => {
+  if (!props.bankAddress) return
+
+  try {
+    if (data.token === 'ETH') {
+      sendTransaction({
+        to: props.bankAddress,
+        value: parseEther(data.amount)
+      })
+    } else if (data.token === 'USDC') {
+      const amount = BigInt(Number(data.amount) * 1e6)
+      depositAmount.value = data.amount // Store amount for after approval
+
+      const allowance = await readContract(config, {
+        address: USDC_ADDRESS as Address,
+        abi: ERC20ABI,
+        functionName: 'allowance',
+        args: [currentAddress as Address, props.bankAddress]
+      })
+
+      const currentAllowance = allowance ? allowance.toString() : 0n
+      if (Number(currentAllowance) < Number(amount)) {
+        approve({
+          address: USDC_ADDRESS as Address,
+          abi: ERC20ABI,
+          functionName: 'approve',
+          args: [props.bankAddress, amount]
+        })
+      } else {
+        // If already approved, deposit directly
+        await handleUsdcDeposit(data.amount)
+      }
+    }
+  } catch (error) {
+    console.error(error)
+    addErrorToast(`Failed to deposit ${data.token}`)
+  }
+}
+
+const handleUsdcDeposit = async (amount: string) => {
+  if (!props.bankAddress) return
+  const tokenAmount = BigInt(Number(amount) * 1e6)
+
+  writeTokenDeposit({
+    address: props.bankAddress,
+    abi: BankABI,
+    functionName: 'depositToken',
+    args: [USDC_ADDRESS as Address, tokenAmount]
+  })
+}
+
+const transferFromBank = async (to: string, amount: string, description: string, token: string) => {
+  if (!props.bankAddress) return
+
+  try {
+    if (token === NETWORK.currencySymbol) {
+      transfer({
+        address: props.bankAddress,
+        abi: BankABI,
+        functionName: 'transfer',
+        args: [to, parseEther(amount)]
+      })
+    } else if (token === 'USDC') {
+      const tokenAmount = BigInt(Number(amount) * 1e6)
+      transfer({
+        address: props.bankAddress,
+        abi: BankABI,
+        functionName: 'transferToken',
+        args: [USDC_ADDRESS as Address, to, tokenAmount]
+      })
+    }
+  } catch (error) {
+    console.error(error)
+    addErrorToast(`Failed to transfer ${token}`)
+  }
+}
+
+const searchUsers = async (input: { name: string; address: string }) => {
+  try {
+    if (input.address) {
+      const { data } = await useCustomFetch(`user/search?address=${input.address}`)
+        .get()
+        .json<{ users: User[] }>()
+      foundUsers.value = data.value?.users || []
+    }
+  } catch (error) {
+    console.error(error)
+    addErrorToast('Failed to search users')
+  }
+}
+
 // Computed properties
 const formattedUsdcBalance = computed(() =>
   usdcBalance.value ? (Number(usdcBalance.value) / 1e6).toString() : '0'
 )
 
-// Watch for balance errors
+const loadingText = computed(() => {
+  if (isPendingApprove.value) return 'Approving USDC...'
+  if (isConfirmingApprove.value) return 'Confirming USDC approval...'
+  if (tokenDepositLoading.value) return 'Depositing USDC...'
+  if (isConfirmingTokenDeposit.value) return 'Confirming USDC deposit...'
+  if (depositLoading.value) return 'Depositing ETH...'
+  if (isConfirmingDeposit.value) return 'Confirming ETH deposit...'
+  return 'Processing...'
+})
+
+// Watch handlers
 watch(balanceError, () => {
   if (balanceError.value) {
     addErrorToast('Failed to fetch team balance')
@@ -117,12 +312,10 @@ watch(balanceError, () => {
   }
 })
 
-// Watch for balance updates
 watch([teamBalance, usdcBalance], () => {
   emit('balance-updated')
 })
 
-// Watch for bank address changes
 watch(
   () => props.bankAddress,
   async (newAddress) => {
@@ -131,6 +324,40 @@ watch(
     }
   }
 )
+
+watch(isConfirmingDeposit, (newIsConfirming, oldIsConfirming) => {
+  if (!newIsConfirming && oldIsConfirming) {
+    addSuccessToast('ETH deposited successfully')
+    depositModal.value = false
+    fetchBalance()
+  }
+})
+
+watch(isConfirmingTransfer, (newIsConfirming, oldIsConfirming) => {
+  if (!newIsConfirming && oldIsConfirming) {
+    addSuccessToast('Transferred successfully')
+    transferModal.value = false
+    Promise.all([fetchBalance(), fetchUsdcBalance()])
+  }
+})
+
+watch(isConfirmingTokenDeposit, (newIsConfirming, oldIsConfirming) => {
+  if (!newIsConfirming && oldIsConfirming) {
+    addSuccessToast('USDC deposited successfully')
+    depositModal.value = false
+    fetchUsdcBalance()
+    depositAmount.value = '' // Clear stored amount
+  }
+})
+
+watch(isConfirmingApprove, async (newIsConfirming, oldIsConfirming) => {
+  if (!newIsConfirming && oldIsConfirming) {
+    addSuccessToast('Token approved successfully')
+    if (depositAmount.value) {
+      await handleUsdcDeposit(depositAmount.value)
+    }
+  }
+})
 
 // Expose methods and data for parent component
 defineExpose({
